# LMS/DTOs/Tokens.cs
namespace LMS.DTOs
{
    public class TokenRequest
    {
        public string accessToken { get; set; }
        public string refreshToken { get; set; }
    }

    public class TokenResponse
    {
        public string accessToken { get; set; }
        public string refreshToken { get; set; }
    }

}